[#159958736] Replace code from temporary files
