+++ src/_components/Assets/AddAssetContainer.jsx
@@ -2,11 +2,13 @@ import _ from 'lodash';
 import PropTypes from 'prop-types';
 import * as React from 'react';
 import { connect } from 'react-redux';
+import { Divider, Grid, Header, Icon, Step } from 'semantic-ui-react';
 import { ToastMessage } from '../../_utils/ToastMessage';
 
-import AddAssetComponent from '../../components/Assets/AddAssetComponent';
 import FilterAssetComponent from '../../components/Assets/FilterAssetComponent';
 import SpecsComponent from '../../components/Assets/SpecsComponent';
+import NavbarComponent from '../../components/NavBarComponent';
+import LoaderComponent from '../../components/LoaderComponent';
 
 import { loadCategoriesDropdown } from '../../_actions/category.actions';
 import { loadSubCategoriesDropdown } from '../../_actions/subcategory.actions';
@@ -23,7 +25,7 @@ import {
   filterModelNumbers
 } from '../../_utils/filterDropdowns';
 
-class AddAssetContainer extends React.Component {
+class AddAssetContainer2 extends React.Component {
   state = {
     filteredSubCategories: [],
     filteredAssetTypes: [],
@@ -36,8 +38,7 @@ class AddAssetContainer extends React.Component {
     modelNumber: '',
     serialNumber: '',
     assetTag: '',
-    saveButtonState: false,
-    page: 0,
+    step: 1,
     specs: {
       year: '',
       processorType: '',
@@ -80,7 +81,6 @@ class AddAssetContainer extends React.Component {
       }
 
       nextProps.resetToastMessageContent();
-      nextProps.toggleModal();
 
       return {
         filteredSubCategories: [],
@@ -95,16 +95,11 @@ class AddAssetContainer extends React.Component {
     return null;
   }
 
-  pageValidator = () => {
-    if (this.state.page === 0) {
-      return (_.isEmpty(this.state.selectedAssetMake));
-    }
-    return (
-      _.isEmpty(this.state.modelNumber) ||
-      _.isEmpty(this.state.serialNumber) ||
-      _.isEmpty(this.state.assetTag)
-    );
-  }
+  stepValidator = () => (
+    _.isEmpty(this.state.modelNumber) ||
+    _.isEmpty(this.state.serialNumber) ||
+    _.isEmpty(this.state.assetTag)
+  );
 
   handleDropdownChanges = (event, data) => {
     event.stopPropagation();
@@ -157,21 +152,17 @@ class AddAssetContainer extends React.Component {
     this.setState({ assetTag: event.target.value });
   };
 
-  onChangeButtonState = () => {
-    this.setState({ saveButtonState: !this.state.saveButtonState });
+  onNextClicked = () => {
+    this.setState(({ step }) => ({ step: step + 1 }));
   };
 
-  onNextClicked = () => {
-    this.setState(({ page }) => ({ page: page + 1 }));
-  }
-
   goBack = () => {
-    this.setState(({ page }) => ({ page: page - 1 }));
-  }
+    this.setState(({ step }) => ({ step: step - 1 }));
+  };
 
   pickRadioValuesFromSpecsComponent = (data) => {
     this.setState({ specs: data });
-  }
+  };
 
   onCreateAsset = () => {
     const newAssetDetails = {
@@ -179,7 +170,8 @@ class AddAssetContainer extends React.Component {
       serial_number: this.state.serialNumber,
       model_number: this.state.modelNumber
     };
-    if (this.state.page === 2) {
+
+    if (_.isEmpty(this.state.specs)) {
       newAssetDetails.year = this.state.specs.year;
       newAssetDetails.processor_type = this.state.specs.processorType;
       newAssetDetails.processor_speed = this.state.specs.processorSpeed;
@@ -187,67 +179,119 @@ class AddAssetContainer extends React.Component {
       newAssetDetails.storage = this.state.specs.storage;
       newAssetDetails.memory = this.state.specs.memory;
     }
-    this.props.createAsset(newAssetDetails);
+
+    this.props.createAsset(newAssetDetails)
+      .then(() => this.props.history.push('/assets'));
   };
 
   render() {
-    const isDisabled = this.pageValidator();
+    const { step } = this.state;
+    const isDisabled = this.stepValidator();
+    let stepContent;
+    let step1Active;
+    let step2Active;
+    const acceptableAssetTypes = [
+      'Macbook', 'Monitor', 'Monitors', 'Phone', 'Tablet'
+    ];
 
-    if (this.state.page === 1) {
+    const isAssetSpecsAvailable = () =>
+      acceptableAssetTypes.indexOf(this.state.selectedAssetType) > -1;
+
+    if (Object.values(this.props.isLoadingState)
+      .find(loading => loading)) {
       return (
-        <AddAssetComponent
-          {...this.props}
-          onSelectModelNumber={this.onSelectModelNumber}
-          onAddSerialNumber={this.onAddSerialNumber}
-          onAddAssetTag={this.onAddAssetTag}
-          onCreateAsset={this.onCreateAsset}
-          goBack={this.goBack}
-          onNextClicked={this.onNextClicked}
-          filteredModelNumbers={this.state.filteredModelNumbers}
-          filteredAssetTypes={this.state.filteredAssetTypes}
-          modelNumber={this.state.modelNumber}
-          serialNumber={this.state.serialNumber}
-          assetTag={this.state.assetTag}
-          selectedAssetType={this.state.selectedAssetType}
-          buttonState={this.state.saveButtonState}
-          onChangeButtonState={this.onChangeButtonState}
-          isDisabled={isDisabled}
-        />
-      );
-    } else if (this.state.page === 2) {
-      return (
-        <SpecsComponent
-          {...this.props}
-          specs={this.state.specs}
-          goBack={this.goBack}
-          buttonState={this.state.saveButtonState}
-          onChangeButtonState={this.onChangeButtonState}
-          page={this.state.page}
-          onCreateAsset={this.onCreateAsset}
-          pickRadioValuesFromSpecsComponent={this.pickRadioValuesFromSpecsComponent}
-        />
+        <NavbarComponent>
+          <LoaderComponent />
+        </NavbarComponent>
       );
     }
+
+    if (step === 1) {
+      step1Active = step === 1;
+
+      stepContent = (
+        <div className="add-asset__device-info">
+          <FilterAssetComponent
+            {...this.props}
+            handleDropdownChanges={this.handleDropdownChanges}
+            filteredSubCategories={this.state.filteredSubCategories}
+            filteredAssetTypes={this.state.filteredAssetTypes}
+            filteredAssetMakes={this.state.filteredAssetMakes}
+            filteredModelNumbers={this.state.filteredModelNumbers}
+            onSelectModelNumber={this.onSelectModelNumber}
+            onAddSerialNumber={this.onAddSerialNumber}
+            onAddAssetTag={this.onAddAssetTag}
+            selectedCategory={this.state.selectedCategory}
+            selectedSubcategory={this.state.selectedSubcategory}
+            selectedAssetType={this.state.selectedAssetType}
+            selectedAssetMake={this.state.selectedAssetMake}
+            modelNumber={this.state.modelNumber}
+            serialNumber={this.state.serialNumber}
+            assetTag={this.state.assetTag}
+            isDisabled={isDisabled}
+            onNextClicked={this.onNextClicked}
+            isAssetSpecsAvailable={isAssetSpecsAvailable()}
+            onChangeButtonState={this.onChangeButtonState}
+            buttonLoading={this.props.buttonLoading}
+            onCreateAsset={this.onCreateAsset}
+          />
+        </div>
+      );
+    } else if (step === 2) {
+      step2Active = step === 2;
+
+      stepContent = (
+        <div className="add-asset__device-specs">
+          <SpecsComponent
+            specs={this.state.specs}
+            goBack={this.goBack}
+            buttonLoading={this.props.buttonLoading}
+            onChangeButtonState={this.onChangeButtonState}
+            onCreateAsset={this.onCreateAsset}
+            pickRadioValuesFromSpecsComponent={this.pickRadioValuesFromSpecsComponent}
+          />
+        </div>
+      );
+    }
+
     return (
-      <FilterAssetComponent
-        {...this.props}
-        handleDropdownChanges={this.handleDropdownChanges}
-        filteredSubCategories={this.state.filteredSubCategories}
-        filteredAssetTypes={this.state.filteredAssetTypes}
-        filteredAssetMakes={this.state.filteredAssetMakes}
-        page={this.state.page}
-        selectedCategory={this.state.selectedCategory}
-        selectedSubcategory={this.state.selectedSubcategory}
-        selectedAssetType={this.state.selectedAssetType}
-        selectedAssetMake={this.state.selectedAssetMake}
-        isDisabled={isDisabled}
-        onNextClicked={this.onNextClicked}
-      />
+      <NavbarComponent>
+        <div className="add-asset">
+          <div id="page-heading-section">
+            <Header as="h1" id="page-headings" floated="left" content="Add Asset" />
+            <Divider id="assets-divider" />
+          </div>
+
+          <Grid centered columns={2}>
+            <Grid.Column>
+              <Step.Group size="small" widths={2}>
+                <Step active={step1Active} completed={!step1Active}>
+                  <Icon name="laptop" />
+                  <Step.Content>
+                    <Step.Title>Device Info</Step.Title>
+                    <Step.Description>Identify your device</Step.Description>
+                  </Step.Content>
+                </Step>
+
+                <Step active={step2Active} disabled={step1Active}>
+                  <Icon name="list ul" />
+                  <Step.Content>
+                    <Step.Title>Device Specs</Step.Title>
+                    <Step.Description>Enter device specifications</Step.Description>
+                  </Step.Content>
+                </Step>
+              </Step.Group>
+
+              {stepContent}
+            </Grid.Column>
+          </Grid>
+        </div>
+      </NavbarComponent>
     );
   }
 }
 
-AddAssetContainer.propTypes = {
+AddAssetContainer2.propTypes = {
   categories: PropTypes.array,
   subcategories: PropTypes.array,
   assetTypes: PropTypes.array,
@@ -262,10 +306,13 @@ AddAssetContainer.propTypes = {
   loadAssetMakesDropdown: PropTypes.func.isRequired,
   loadModelNumbers: PropTypes.func.isRequired,
   createAsset: PropTypes.func.isRequired,
-  resetToastMessageContent: PropTypes.func.isRequired
+  resetToastMessageContent: PropTypes.func.isRequired,
+  isLoadingState: PropTypes.object.isRequired,
+  buttonLoading: PropTypes.bool.isRequired,
+  history: PropTypes.object
 };
 
-AddAssetContainer.defaultProps = {
+AddAssetContainer2.defaultProps = {
   categories: [],
   subcategories: [],
   assetTypes: [],
@@ -275,7 +322,8 @@ AddAssetContainer.defaultProps = {
   toastMessageContent: {
     type: '',
     message: ''
-  }
+  },
+  history: {}
 };
 
 const mapStateToProps = ({
@@ -294,6 +342,7 @@ const mapStateToProps = ({
   modelNumbers: modelNumbersList,
   assets: assets.assetsList,
   toastMessageContent: toastMessage,
+  buttonLoading: assets.isLoading,
   isLoadingState: {
     isLoadingCategories: categoriesList.isLoading,
     isLoadingSubcategories: subcategoriesList.isLoading,
@@ -311,4 +360,4 @@ export default connect(mapStateToProps, {
   loadModelNumbers,
   createAsset,
   resetToastMessageContent
-})(AddAssetContainer);
+})(AddAssetContainer2);
