+++ src/_components/RoutesComponent.jsx
@@ -18,7 +18,7 @@ import AssetMakes from '../components/AssetMake/AssetMakeComponent';
 import User from './User/UserContainer';
 import UserDetail from './User/UserDetailContainer';
 import AssetSpecs from '../components/AssetSpecs/AssetSpecsComponent';
-import AddAsset2 from './Assets/AddAssetContainer2';
+import AddAsset from './Assets/AddAssetContainer';
 
 class RoutesComponent extends Component {
   checkAuthentication = () => !!(localStorage.getItem('art-prod-web-token'));
@@ -117,8 +117,8 @@ class RoutesComponent extends Component {
           <Authenticate
             exact
             isAuthenticated={this.checkAuthentication()}
-            path="/add-asset"
-            component={AddAsset2}
+            path="/assets/create"
+            component={AddAsset}
           />
           <Route exact path="/" component={LoginComponent} />
           <Route path="*" component={PageNotFound} />
