+++ src/_constants/index.js
@@ -27,6 +27,7 @@ const CONSTANTS = KeyMirror({
   LOAD_ASSET_MAKES_SUCCESS: true,
   LOAD_ASSET_MAKES_FAILURE: true,
   LOAD_ASSET_MODEL_NUMBERS: true,
+  CREATE_ASSET_REQUEST: true,
   CREATE_ASSET_SUCCESS: true,
   CREATE_ASSET_FAIL: true,
   LOAD_FEEDBACK_SUCCESS: true,
