+++ src/_reducers/assets.reducer.js
@@ -2,6 +2,7 @@ import constants from '../_constants';
 import initialState from './initialState';
 
 const {
+  CREATE_ASSET_REQUEST,
   CREATE_ASSET_SUCCESS,
   CREATE_ASSET_FAIL,
   LOAD_ASSETS_SUCCESS,
@@ -12,6 +13,12 @@ const {
 
 export default (state = initialState.assets, action) => {
   switch (action.type) {
+    case CREATE_ASSET_REQUEST:
+      return {
+        ...state,
+        hasError: false,
+        isLoading: true
+      };
     case CREATE_ASSET_SUCCESS:
       return {
         ...state,
